renames engine.plan to plan.deployment and moves associated modules into own submodule
