+++ org.opentosca.container.control/src/main/java/org/opentosca/container/control/impl/OpenToscaControlServiceImpl.java
@@ -23,7 +23,8 @@ import org.opentosca.container.core.next.model.DeploymentProcessState;
 import org.opentosca.container.core.service.CsarStorageService;
 import org.opentosca.container.core.service.DeploymentTracker;
 import org.opentosca.container.core.service.IPlanInvocationEngine;
-import org.opentosca.container.engine.plan.IPlanEngineService;
+
+import org.opentosca.container.plan.deployment.core.IPlanEngineService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
