+++ org.opentosca.container.control/src/main/resources/spring/context-control.xml
@@ -14,7 +14,7 @@
 
   <context:component-scan base-package="org.opentosca.container.core.impl.service"/>
   <context:component-scan base-package="org.opentosca.container.core.plan"/>
-  <context:component-scan base-package="org.opentosca.container.engine.plan.impl"/>
+  <context:component-scan base-package="org.opentosca.container.plan.deployment.core"/>
 
   <bean class="org.opentosca.container.control.impl.OpenToscaControlServiceImpl" />
 </beans>
