--- org.opentosca.container.engine.plan/pom.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-         xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.opentosca</groupId>
-        <artifactId>container</artifactId>
-        <version>3.0.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>org.opentosca.container.engine.plan</artifactId>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.github.opentosca.winery</groupId>
-            <artifactId>org.eclipse.winery.model.tosca.canonical</artifactId>
-            <version>ustutt-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.opentosca</groupId>
-            <artifactId>org.opentosca.container.core</artifactId>
-            <version>${project.version}</version>
-            <scope>compile</scope>
-        </dependency>
-    </dependencies>
-</project>
