+++ org.opentosca.container.plan.deployment/org.opentosca.container.plan.deployment.plugin.bpel/src/main/java/org/apache/ode/schemas/dd/_2007/_03/TEnableSharing.java
