+++ org.opentosca.container.plan.deployment/pom.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.opentosca</groupId>
+    <artifactId>container</artifactId>
+    <version>3.0.0-SNAPSHOT</version>
+  </parent>
+
+  <artifactId>org.opentosca.container.plan.deployment</artifactId>
+  <packaging>pom</packaging>
+
+  <modules>
+    <module>org.opentosca.container.plan.deployment.core</module>
+    <module>org.opentosca.container.plan.deployment.plugin.bpel</module>
+    <module>org.opentosca.container.plan.deployment.plugin.camunda</module>
+  </modules>
+</project>
