+++ org.opentosca.container.reporting/pom.xml
@@ -231,7 +231,7 @@
 
       <dependency>
           <groupId>org.opentosca</groupId>
-          <artifactId>org.opentosca.container.engine.plan.plugin.camunda</artifactId>
+          <artifactId>org.opentosca.container.plan.deployment.plugin.camunda</artifactId>
           <version>${project.version}</version>
           <exclusions>
               <exclusion>
@@ -242,7 +242,7 @@
       </dependency>
       <dependency>
           <groupId>org.opentosca</groupId>
-          <artifactId>org.opentosca.container.engine.plan.plugin.bpel</artifactId>
+          <artifactId>org.opentosca.container.plan.deployment.plugin.bpel</artifactId>
           <version>${project.version}</version>
           <exclusions>
               <exclusion>
